Return value of Net.nm_dot2bit() could be a float. Fixed by casting to an int before returning.
git-svn-id: https://svn.electricmonk.nl/svn/dataq/trunk@63 5e763a1a-bcdf-427f-adeb-eebfbc5de0a3

+++ src/dataq.py
@@ -226,7 +226,7 @@ class Net:
 		
 		l = Net.dot2long(dot)
 
-		return(32 - math.log( (1 << 32L) - l, 2))
+		return(int(32 - math.log( (1 << 32L) - l, 2)))
 
 	def nm_bit2dot(bit):
 		"""
